+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Changelog
 
-### 15/11/2022 - React Session #6
+### 15/11/2022 - React Review # 8
 
 #### Changes applied
 
@@ -12,3 +12,13 @@
 ### Changes Planned
 
 -   [TODO] Create new page Profile.
+
+### 17/11/2022 - React Session # 9
+
+#### Changes Planned
+
+-   [DONE][0.3h] Create a contextAPI module to be used as a hook to other components.
+-   [WIP][0.1h] Consume the ContextAPI in a component
+-   [WIP][0.5h] Display the data stored in the ContextAPI hook.
+-   [TODO][0.3h] Querying and storing data retrieved from an API.
+-   [TODO][1.0h] Create new page Profile.